Separate out tag scanning spec
