+++ core/src/test/scala/akka/persistence/cassandra/EventsByTagRestartSpec.scala
@@ -7,12 +7,10 @@ package akka.persistence.cassandra
 import java.time.format.DateTimeFormatter
 import java.time.{ LocalDateTime, ZoneOffset }
 
-import akka.NotUsed
-import akka.persistence.cassandra.TestTaggingActor.{ Ack, Crash, Stop }
+import akka.persistence.cassandra.TestTaggingActor.{ Ack, Stop }
 import akka.persistence.cassandra.query.scaladsl.CassandraReadJournal
 import akka.persistence.query.{ EventEnvelope, NoOffset, PersistenceQuery }
 import akka.stream.ActorMaterializer
-import akka.stream.scaladsl.Source
 import akka.stream.testkit.scaladsl.TestSink
 import akka.testkit.TestProbe
 import com.typesafe.config.ConfigFactory
@@ -82,32 +80,6 @@ class EventsByTagRestartSpec extends CassandraSpec(EventsByTagRestartSpec.config
       }
       tagProbe.expectNoMessage(waitTime)
       tagProbe.cancel()
-    }
-
-    "continue tag sequence nrs for actor crashing" in {
-      // crash the actor 250 times, persist 5 events each time
-      val crashEvery = 5
-      val crashNr = 250
-      val msgs = crashEvery * crashNr
-      val p2 = system.actorOf(TestTaggingActor.props("p2", Set("blue")))
-      (1 to msgs).foreach { cn =>
-        if (cn % crashEvery == 0) {
-          p2 ! Crash
-        }
-        val msg = s"msg $cn"
-        p2 ! msg
-        expectMsg(Ack)
-      }
-      val blueTags: Source[EventEnvelope, NotUsed] = queryJournal.eventsByTag(tag = "blue", offset = NoOffset)
-      val tagProbe = blueTags.runWith(TestSink.probe[EventEnvelope](system))
-      (1L to msgs).foreach { m =>
-        val expected = s"msg $m"
-        tagProbe.request(1)
-        tagProbe.expectNext().event shouldEqual expected
-      }
-      tagProbe.expectNoMessage(250.millis)
-      tagProbe.cancel()
-
     }
   }
 }
